// prisma/generated/type-graphql/models/Checklist.ts
import * as TypeGraphQL from "type-graphql";
import * as GraphQLScalars from "graphql-scalars";
import { Prisma } from "@prisma/client";
import { DecimalJSScalar } from "../scalars";
import { Trainings } from "../models/Trainings";
import { ChecklistCount } from "../resolvers/outputs/ChecklistCount";

@TypeGraphQL.ObjectType("Checklist", {
  isAbstract: true
})
export class Checklist {
  @TypeGraphQL.Field(_type => GraphQLScalars.BigIntResolver, {
    nullable: false
  })
  id!: bigint;

  @TypeGraphQL.Field(_type => Date, {
    nullable: true
  })
  created_at?: Date | null;

  @TypeGraphQL.Field(_type => String, {
    nullable: true
  })
  username?: string | null;

  @TypeGraphQL.Field(_type => String, {
    nullable: true
  })
  location?: string | null;

  @TypeGraphQL.Field(_type => String, {
    nullable: true
  })
  regional_manager?: string | null;

  @TypeGraphQL.Field(_type => String, {
    nullable: true
  })
  manager?: string | null;

  @TypeGraphQL.Field(_type => String, {
    nullable: true
  })
  description?: string | null;

  @TypeGraphQL.Field(_type => String, {
    nullable: true
  })
  ticket?: string | null;

  @TypeGraphQL.Field(_type => String, {
    nullable: true
  })
  project?: string | null;

  @TypeGraphQL.Field(_type => String, {
    nullable: true
  })
  region?: string | null;

  @TypeGraphQL.Field(_type => String, {
    nullable: true
  })
  country?: string | null;

  @TypeGraphQL.Field(_type => String, {
    nullable: true
  })
  media?: string | null;

  @TypeGraphQL.Field(_type => String, {
    nullable: true
  })
  prod_id?: string | null;

  trainings?: Trainings[];

  @TypeGraphQL.Field(_type => ChecklistCount, {
    nullable: true
  })
  _count?: ChecklistCount | null;
}